Fixed leaks in unittests. Moved mongoConnectionPool to separate module
